% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_data.R
\name{close_cycling_data}
\alias{close_cycling_data}
\title{Find closest data from Wang and Dillon 2014 to a specified location within a
time interval.}
\usage{
close_cycling_data(samp_coords = stop("Sample coords c(lon,lat) required."),
  samp_interval = NULL, ...)
}
\arguments{
\item{samp_coords}{The \code{cbind(lon, lat)} of the search origin, in
decimal degrees, negative south and west.}

\item{samp_interval}{Period over which to search for data \code{c(start,end),
"mm/dd/yyyy"}. Default returns all.}

\item{...}{additional parameters to limit search (see \code{close_stations}).}
}
\value{
a subset of the cycling data, with and additional column of distance
  from sample, sorted by distance to the specified location.
}
\description{
Search through the Wang and Dillon cycling dataset, and find the closest
stations to a specified sample site.
}
\examples{
close_cycling_data(samp_coords = c(9.0556, 48.52),
                   samp_interval = c("1/1/1998", "4/7/2007"),
                   radius = 50,
                   n = 2)
}
\seealso{
\code{\link{close_stations}} for constraining search.
}

