% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_data.R
\name{close_noaaisd_data}
\alias{close_noaaisd_data}
\title{Find closest data from the NOAA ISD dataset to a specified location and time.}
\usage{
close_noaaisd_data(samp_coords, samp_interval = NULL, ...)
}
\arguments{
\item{samp_coords}{The \code{cbind(lon, lat)} of the search origin, in decimal
degrees, negative south and west.}

\item{samp_interval}{Period over which to search for data \code{c(start,end),
"mm/dd/yyyy"}. Default returns all.}

\item{...}{additional parameters to limit search (see \code{close_stations}).}
}
\value{
a data frame with \code{usaf} and \code{wban} identifiers for use
  with \code{get_noaaisd_data}.
}
\description{
Search through the Wang & Dillon cycling dataset, and find the closest
stations to a specified sample site.
}
\examples{
close_noaaisd_data(samp_coords = c(9.0556, 48.52),
                   samp_interval = c("1/1/1998", "4/7/2007"),
                   radius = 50,
                   n = 2)
}
\seealso{
\code{\link{close_stations}} for constraining search.
}

