% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_data.R
\name{get_noaaisd_data}
\alias{get_noaaisd_data}
\title{Download and temperature data NOAA ISD and format for spectral analysis.}
\usage{
get_noaaisd_data(usaf = stop("6-digit char USAF id required."),
  wban = stop("5-digit char WBAN id required."), years = NULL)
}
\arguments{
\item{usaf}{A string of the USAF weather station identifier.}

\item{wban}{A string of the WBAN weather station identifier.}

\item{years}{A vector of years to download. Default returns all, missing data
results in a warning.}
}
\value{
A data frame with \code{jday} and \code{temperature}.
}
\description{
A helper script for the \code{rnoaa::isd} function. Download the specified
weather station for the specified years, remove errors and low quality data,
convert to julian day, and concatenate into a single data frame. Temperature
values > 900 are error marks and are removed. Only values which pass all
quality control tests, (quality code "1"), are retained.
}
\examples{
get_noaaisd_data(usaf = "702700",
                 wban = "00489",
                 years = c(2009,2013:2015))
}

