% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_data.R
\name{mmddyyyy_to_numeric}
\alias{mmddyyyy_to_numeric}
\title{mm/dd/yyyy to julian date helper function}
\usage{
mmddyyyy_to_numeric(mmddyyyy_dates)
}
\arguments{
\item{mmddyyyy_dates}{a vector of dates in "mm/dd/yyyy" format.}
}
\value{
a numeric vector of the number of days since 1/1/1960, dates before
  this are negative.
}
\description{
The \pkg{rnoaa} package returns dates in the station list in "mm/dd/yyyy"
format, and we (Wang & Dillon) formatted our dates in a similar way. All date
and spectral calculations work on multi-year Julian dates. This is a small
helper function to convert from "mm/dd/yyyy" to julian day. Because it uses
\code{mdy.date()} from the \pkg{date} package, it returns the number of days
since January 1, 1960. Inputs are coerced to character if necessary.
}
\examples{
mmddyyyy_to_numeric(c("3/5/1954", "9/28/2012"))
}

