% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_data.R
\name{prelim_record_check}
\alias{prelim_record_check}
\title{Preliminary checks for temperature records.}
\usage{
prelim_record_check(record_time)
}
\arguments{
\item{record_time}{A time vector, in fractional julian date.}
}
\value{
A logical data frame with check values.
}
\description{
Check if a weather record is suitable for spectral analysis. Checks are based
on sampling: records must sample at least every six hours on average, be at
least 1.5 years in length, and not have a single gap more than ten percent of
the record length. As suggested by the function name, these checks are
preliminary. The significance of the peaks should also be measured, either by
comparison to a white noise model, as in \pkg{nlts}, or against Monte Carlo
or Chi-square AR1 levels as in redfit, available in \pkg{dplR}.
}
\examples{
weather_record <- get_noaaisd_data(usaf = "702700",
                                   wban = "00489",
                                   years = 2014:2015)
prelim_record_check(weather_record$jday)
}

