% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_data.R
\name{yyyymmdd_to_numeric}
\alias{yyyymmdd_to_numeric}
\title{yyyy-mm-dd to julian date helper function}
\usage{
yyyymmdd_to_numeric(yyyymmdd_dates = stop("Char dates yyyy-mm-dd rqrd."))
}
\arguments{
\item{yyyymmdd_dates}{a vector of dates in "yyyy-mm-dd" format.}
}
\value{
a numeric vector of the number of days since 1/1/1960, dates before
  this are negative.
}
\description{
The \pkg{rnoaa} package returns dates for weather station records in
"yyyy-mm-dd" format. All date and spectral calculations work on multi-year
Julian dates. This is a small helper function to convert from "yyyy-mm-dd" to
julian day. Because it uses \code{mdy.date()} from the \pkg{date} package, it
returns the number of days since January 1, 1960. Inputs are coerced to
character if necessary.
}
\examples{
yyyymmdd_to_numeric(c("1954-3-5", "2012-9-28"))
}

