% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incoming_data.R
\name{prelim_record_check}
\alias{prelim_record_check}
\title{Preliminary checks for temperature records.}
\usage{
prelim_record_check()
}
\arguments{
\item{usaf}{A string of the USAF weather station identifier.}

\item{wban}{A string of the WBAN weather station identifier.}
}
\value{
A data frame with \code{jday} and \code{temperature}.
}
\description{
Preliminary checks for temperature records.
}
\examples{
get_noaaisd_data(usaf = "702700",
                 wban = "00489",
                 years = c(2009,2013:2015))
}

